refactored nfact_pp utils into general utils
