+++ NFACT/NFACT_PP/probtrackx_functions.py
@@ -2,7 +2,7 @@ import os
 import subprocess
 import multiprocessing
 import signal
-from NFACT.NFACT_PP.nfactpp_utils import write_to_file, date_for_filename
+from NFACT.utils.utils import write_to_file, date_for_filename
 from NFACT.utils.utils import colours, error_and_exit
